Tests/Dbus: Move DBus-related files to the tests/dbus directory
This patch relocates the dbus interface for the unit tests to the tests
directory. According to the changes, the meson build files are alse
updated as well.
