+++ dbus/meson.build
@@ -1,3 +1,2 @@
 pipeline_dbus_input = files('pipeline-dbus.xml')
 model_dbus_input = files('model-dbus.xml')
-test_dbus_input = files('test-dbus.xml')
