+++ tests/capi/meson.build
@@ -34,9 +34,10 @@ test('unittest_capi_datatype_consistency', unittest_capi_datatype_consistency, e
 if get_option('enable-ml-service')
   unittest_capi_service_agent_client = executable('unittest_capi_service_agent_client',
     'unittest_capi_service_agent_client.cc',
-    dependencies: [unittest_common_dep, nns_capi_service_dep, gdbus_gen_header_test_dep],
+    dependencies: [unittest_common_dep, nns_capi_service_dep, gdbus_test_gen_dep],
     install: get_option('install-test'),
-    install_dir: unittest_install_dir
+    install_dir: unittest_install_dir,
+    include_directories: [nns_capi_include, nns_ml_agent_incs],
   )
   test('unittest_capi_service_agent_client', unittest_capi_service_agent_client, env: testenv, timeout: 100)
 endif
